fix(build): eliminate Kotlin compiler warnings across screens, player and extractor call sites
- drop redundant elvis, unsafe-call and always-true condition checks
- defer Json config to single hoisted instances (YouTube.kt, WatchMetadataMergeTest)
- remove dead @OptIn(UnstableApi) annotations
- migrate deprecated Compose APIs: hiltViewModel import, AutoMirrored icons,
  PrimaryTabRow/SecondaryScrollableTabRow, ExposedDropdownMenu menuAnchor,
  SwipeToDismissBox confirmValueChange, composition-local clipboard/lifecycle
- replace deprecated NewPipe field accesses (fps, isVideoOnly, url, content)
- suppress API deprecations without replacements; VibratorManager path on S+
- drop 823 Turkish-only string keys from values-tr
- include home-feed cache migration 26to27 and upload-artifact v7

+++ app/src/main/java/com/omersusin/pitube/ui/components/audio/EqualizerEditor.kt
@@ -293,7 +293,7 @@ private fun logToFreq(norm: Float): Double {
     val minLog = kotlin.math.log10(20.0)
     val maxLog = kotlin.math.log10(20000.0)
     val logFreq = minLog + (norm * (maxLog - minLog))
-    return 10.0.pow(logFreq.toDouble())
+    return 10.0.pow(logFreq)
 }
 
 // Editing scaffold shown when the active profile has no bands (e.g. Flat).
